StartDBInstance parameter change.

+++ aliyun-java-sdk-rds/src/main/java/com/aliyuncs/rds/model/v20140815/DescribeVSwitchesRequest.java
@@ -27,11 +27,11 @@ public class DescribeVSwitchesRequest extends RpcAcsRequest<DescribeVSwitchesRes
 
 	private Long resourceOwnerId;
 
-	private Long pageNumber;
+	private Integer pageNumber;
 
 	private String securityToken;
 
-	private Long pageSize;
+	private Integer pageSize;
 
 	private String dedicatedHostGroupId;
 
@@ -64,11 +64,11 @@ public class DescribeVSwitchesRequest extends RpcAcsRequest<DescribeVSwitchesRes
 		}
 	}
 
-	public Long getPageNumber() {
+	public Integer getPageNumber() {
 		return this.pageNumber;
 	}
 
-	public void setPageNumber(Long pageNumber) {
+	public void setPageNumber(Integer pageNumber) {
 		this.pageNumber = pageNumber;
 		if(pageNumber != null){
 			putQueryParameter("PageNumber", pageNumber.toString());
@@ -86,11 +86,11 @@ public class DescribeVSwitchesRequest extends RpcAcsRequest<DescribeVSwitchesRes
 		}
 	}
 
-	public Long getPageSize() {
+	public Integer getPageSize() {
 		return this.pageSize;
 	}
 
-	public void setPageSize(Long pageSize) {
+	public void setPageSize(Integer pageSize) {
 		this.pageSize = pageSize;
 		if(pageSize != null){
 			putQueryParameter("PageSize", pageSize.toString());
